refactor: move genserver to separate file; merge worker and function
Move Worker.Worker GenServer behaviour to a separate file; move Worker.Function struct to Worker.Worker file. Add docs for Worker.Function.
