+++ lib/application.ex
@@ -17,10 +17,12 @@
 #
 
 defmodule WorkerApp do
+  @moduledoc false
+
   use Application
 
   def start(_type, _args) do
-    children = [{Worker.Updater, []}, {Worker.Worker, []}]
+    children = [{Worker.Updater, []}, {Worker.Server, []}]
     Supervisor.start_link(children, strategy: :rest_for_one)
   end
 end
