+++ lib/worker/worker.ex
@@ -16,11 +16,27 @@
 # under the License.
 #
 
+defmodule Worker.Function do
+  @moduledoc """
+    Function struct, passed to Fn.
+
+    ## Fields
+      - name: function name
+      - image: base Docker image for the function's container
+      - archive: tarball containing the function's code, will be copied into container
+      - main_file: path of the function's main file inside the container
+  """
+  @enforce_keys [:name, :image, :archive]
+  defstruct [:name, :image, :archive, :main_file]
+end
+
+
 defmodule Worker.Worker do
-  use GenServer, restart: :permanent
+  @moduledoc """
+
+  """
   alias Worker.Fn
 
-  # Auxiliary functions
   defp get_docker_host() do
     default = "unix:///var/run/docker.sock"
     docker_socket = System.get_env("DOCKER_HOST", default)
@@ -114,48 +130,4 @@ defmodule Worker.Worker do
         {:error, err}
     end
   end
-
-
-  # GenServer behaviour
-
-  def start_link(args) do
-    GenServer.start_link(__MODULE__, args, name: :worker)
-  end
-
-  @impl true
-  def init(_args) do
-    # Process.flag(:trap_exit, true)
-    IO.puts("worker running")
-    {:ok, nil}
-  end
-
-  @impl true
-  def handle_call({:prepare, function}, from, _state) do
-    spawn(__MODULE__, :prepare_container, [function, from])
-    {:noreply, nil}
-  end
-
-  @impl true
-  def handle_call({:run, function}, from, _state) do
-    spawn(__MODULE__, :run_function, [function, from])
-    {:noreply, nil}
-  end
-
-  @impl true
-  def handle_call({:invoke, function}, from, _state) do
-    spawn(__MODULE__, :invoke_function, [function, from])
-    {:noreply, nil}
-  end
-
-  @impl true
-  def handle_call({:cleanup, function}, from, _state) do
-    spawn(__MODULE__, :cleanup, [function, from])
-    {:noreply, nil}
-  end
-
-  # def pipeline(_args) do
-  #   prepare_container("funless-node-container", "node:lts-alpine", "js/hello.tar.gz", "/opt/index.js")
-  #   run_function("funless-node-container")
-  #   cleanup("funless-node-container")
-  # end
 end
